--- Movie List/Movie.swift
@@ -1,29 +1,0 @@
-//
-//  Movie.swift
-//  Movie List
-//
-//  Created by Hannah Bain on 4/18/20.
-//  Copyright © 2020 Lambda School. All rights reserved.
-//
-
-import Foundation
-
-class Movie {
-    var movieName: String
-    var seen: Bool
-    
-    init(movieName: String, seen: Bool) {
-        self.movieName = movieName
-        self.seen = seen
-    }
-    func movieSeen() {
-            if seen == true {
-
-            }
-        }
-    }
-
-
-    
-
-
